refactor(FE): `useAttackFakemon` hook
