+++ frontend/src/components/Battle/Actions.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Card from "react-bootstrap/Card";
 import { useParams } from "react-router-dom";
-import { useAttackFakemon } from "../../hooks/mutations";
+import useAttackFakemon from "../../hooks/useAttackFakemon";
 import useCurrentBattle from "../../hooks/useCurrentBattle";
 import useEndBattle from "../../hooks/useEndBattle";
 import useFakemonsByUser from "../../hooks/useFakemonsByUser";
